Fixed silly checkpoint bug
- The log scan of a decoupled checkpoint must cover the range between
the current chkpt_begin and the last one, so that the complete history
of log records is captured in the latest checkpoint. For that to happen,
the system must first write out a begin log record *and then* start
scanning. Previous code was only inserting a begin after scanning, which
means all log records inserted during the scan were ignored. Really
dumb.