Determining durable_lsn at startup using a backward scan
- Before performing log analysis, the current log file must be
initialized by determining the last valid log record in it. This used to
be in the peek() method, but it used the chk file -- which was removed
recently. This means that a full forward scan was required. New
mechanism scans from the end of the file backwards until a vlaid log
record is found.