Changed log fetch mechanism to copy directly into caller's buffer
- Internal read buffer used for alignment is now inside each partition_t
object. Just like with the partition lock in log_core, a latch was
introduced so that readers on the same partition serialize. The method
release_read() must be called when caller has copied the fetched log
records.
- Also changed the buffer priming mechanism, which is simpler now
because it reuses the partition_t::read method
- Also adapted SPR mechanism, which must read into a buffer of at least
BLOCK_SIZE. It now uses an array of offsets within the buffer to locate
individual log records to be replayed.